loading products

+++ app/src/main/java/com/mercadolibre/mercadopago/domain/model/ProductsModel.kt
@@ -1,7 +1,8 @@
 package com.mercadolibre.mercadopago.domain.model
 
+
 data class ProductsModel(
     val title : String = "",
-    val prace : Int,
+    val prace : Double,
     val image: String
 )
